Adds E2E tests for evaluating recording rules.

This commit adds an E2E test that evaluates 2 samples of a constant
timestamp and checks for the recording metric name and the new sample
in the database.

+++ pkg/promql/engine_test.go
@@ -776,7 +776,7 @@ load 10s
 	cases := []struct {
 		Query               string
 		SkipMaxCheck        bool
-		TotalSamples        int
+		TotalSamples        int64
 		TotalSamplesPerStep stats.TotalSamplesPerStep
 		PeakSamples         int
 		Start               time.Time
